Remove InnerMsg stuff, fix instead by adding another parameter to `updateField` and `selectField`
The problem is: let's say we want to update the model of "Field", a field that is somewhere in the left-hand subtree of "AndThen", which is an AndThen field. AndThen is itself somewhere in the subtree of another field, which we'll call Form.

After we've updated Field, we need to go back up the tree to AndThen, and run its `submit` function - this is because we may need to update the contents of AndThen's right-hand subtree based on the new value of its left-hand subtree.

However, if we only supply Field as a parameter to the `updateField` function, then this is impossible. Field does not contain any of the `update` or `submit` functions for nodes higher up the tree than itself.

The solution is to provide both Form and Field to `updateField`, which can then use Form (which contains _all_ the necessary update and submit functions) to perform the update. Field is now only used to get the Id needed to dispatch the Msg.

It's a less concise and elegant API, but arguably a bit more consistent - `Yafl.update`, `Yafl.submit` etc all take Form as their first argument already, and now `Yafl.updateField` and `Yafl.selectField` (etc) do too.