Try out alternative API for `choice` and `option` in examples/Main.elm
Pros: it has a much simpler implementation than the current choice and option (it's implementable entirely based on existing combinators (map and andThen). It allows the user to pick their own "select an option" field, instead of forcing them to use the built-in one (a set of radio buttons). If adopted, it would not only replace the existing complex implementation, it would also  let us get rid of the entire `Sum` branch/variant of the `Node` type, as well as the special handling for OptionSelected messages. It might also enable us to remove dependencies on elm/html, allowing the library to be used with other UI systems such as elm-ui.

Cons: it's a slightly more complex, and less type-safe API. Also it would need testing to make sure it doesn't do anything weird e.g. with validation.